data-science-application/poetry-ai/pyproject.toml
[project]
name = "poetry-ai-assistant"
version = "1.0.0"
description = "AI-powered poetry analysis, classification, and recommendation service"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Poetry AI Team"}
]
keywords = ["poetry", "ai", "nlp", "classification", "recommendation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.6",
    "langchain-pinecone>=0.0.3",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.20",
    "google-generativeai>=0.3.0",
    "pinecone-client>=3.0.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "duckduckgo-search>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]
streamlit = [
    "streamlit>=1.29.0",
]



[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
